pyproject.toml
[project]
name = "eavesdrop"
version = "0.0.0"
description = "…"
authors = [{name = "shyndman", email = "scotty.hyndman@gmail.com"}]
requires-python = ">=3.12"
dependencies = [
  # These are installed via other means (see Dockerfile)
  # "tokenizers==0.20.3",
  # "transformers[torch]==4.54",
  # "faster-whisper==1.1.1",
  # "openai-whisper==20240930",
  # "ctranslate2 @ file:///~/dev/github.com/ctranslate2",
  # Server dependencies
  "netifaces>=0.11.0",
  "numba",
  "numpy<2",
  "onnxruntime==1.17.0", # Used for VAD
  "psutil",
  "sentencepiece",
  "structlog>=25.4.0",
  "websockets",
  # Audio processing (commented in requirements but including for documentation)
  "kaldialign",
  "soundfile",
  "jiwer",
  "evaluate",
  "websocket-client>=1.8.0",
  "pyyaml>=6.0",
]
license = {text = "MIT"}

[project.urls]
repository = "https://github.com/shyndman/eavesdrop"
"Bug Tracker" = "https://github.com/shyndman/eavesdrop/issues"

[project.scripts]
eavesdrop = "eavesdrop.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/eavesdrop"]

[dependency-groups]
dev = ["pytest>=8.3.5"]

[tool.pyright]
include = ["src", "tests"]
ignore = ["**/.venv"]
defineConstant = {DEBUG = true}
reportMissingImports = "error"
deprecateTypingAliases = true
reportDeprecated = "information"
pythonPlatform = "Linux"

[tool.ruff]
line-length = 100
target-version = "py311"
indent-width = 2

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = []

[tool.ruff.lint.isort]
known-first-party = ["eavesdrop"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "auto"
